refactor(lexer): replace token kind constants with a TokenKind enum (#24931)
Review feedback was that the token kinds should be a real enum rather than a module of u8 constants. The values were listed twice, once as the constants and again in the token_kind_name match, so they now come from one declaration, and tok_kinds and the new Lexer::kinds hand out &[TokenKind] instead of raw bytes.

The discriminants do not move. The pipeline computes kinds arithmetically and blends them in SIMD registers, so it keeps working on u8. Only the aliases in pipeline and the compile-time tables in opmap are re-expressed in terms of the enum, and those fold away at compile time. Reading the bytes back as TokenKind assumes every kind written is a declared discriminant, which a debug assertion checks on every accessor. Output is identical across both corpora and an interleaved A/B against the branch below lands inside the noise floor.

+++ crates/oxc_lexer/src/pipeline/classify.rs
@@ -180,7 +180,7 @@ const fn cls_table(ts: bool) -> [u16; 256] {
     let mut punct = [PUNCT1_KIND_UNKNOWN; 256];
     let mut i = 0;
     while i < PUNCT1_NKNOWN {
-        punct[PUNCT1[i].0 as usize] = PUNCT1[i].1;
+        punct[PUNCT1[i].0 as usize] = PUNCT1[i].1 as u8;
         i += 1;
     }
     let mut t = [0u16; 256];
